services/wiki_service/pyproject.toml
[project]
name = "wiki-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "aiohttp>=3.13",
  "grpcio>=1.78",
  "scpedia-protos @ file:../../libs/scpedia_protos",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
wiki-service = "wiki_service.main:run"